Ignore update-status hook, check leaders after IDLE_CONFIRMATION loop
These two changes are intended to reduce wait time on large
environments.

~Recent juju (1.24+?) regularly runs housekeeping hooks, the most common
of which is update-status.  On a large environment, there is almost
always an update-status run in the last 15 seconds, but juju-wait should
not block on this.

Run the leaders block after the IDLE_CONFIRMATION block, as checking
leadership on an environment with many services takes a significant
amount of time, and is the least likely to set ready = False and begin
the whole process again.  Running this last will make it so it's
(usually) only ever run once, at the very end.